requirements.txt
python-dotenv
requests
beautifulsoup4
schedule
lxml